+++ src/service/agent/BUILD
@@ -32,6 +32,7 @@ osmo_py_library(
         requirement("fastapi"),
         requirement("uvicorn"),
         "//src/utils:backend_messages",
+        "//src/utils:ssl_config",
         "//src/utils/metrics",
         "//src/utils/progress_check:progress_check_lib",
         "//src/service/core/auth",
