+++ src/service/logger/BUILD
@@ -49,6 +49,7 @@ osmo_py_library(
         "//src/service/core/auth",
         "//src/service/core/config:configmap_loader_lib",
         "//src/service/core/workflow",
+        "//src/utils:ssl_config",
         "//src/utils:static_config",
         "//src/utils/job",
     ],
